refactor(core): split transcription/TTS types into types::transcription submodule
First step of the `core/src/types.rs` kitchen-sink split (1675 lines → many
focused submodules). Moves `TranscriptionProviderKind`, `TranscriptionConfig`,
`TtsProviderKind`, and `TtsConfig` out of the flat root namespace into
`assistant_core::types::transcription::*`, dropping the flat re-export from
`lib.rs`.

Why no flat re-export: the lib.rs `pub use types::{50 types}` block is what
makes types.rs a kitchen sink — adding new types there is frictionless, so
they keep accreting. Forcing consumers to import from the canonical
submodule path (`use assistant_core::types::transcription::TtsConfig;`)
makes the module structure visible and gives new types an obvious home.

Call-site changes: 1 file (`crates/transcription/src/lib.rs`) updates its
import. `types.rs` itself imports the relocated types via
`use self::transcription::{...}` for the `AssistantConfig` struct field.

Tests:
- `cargo test -p assistant-core --lib` — 183 pass
- `cargo test -p assistant-transcription --lib` — 36 pass
- `cargo clippy --workspace -- -D warnings` — green

Subsequent PRs will migrate the remaining domains (channels, llm, storage,
mcp, observability, runtime features, conversation, agent config) one at a
time. Each removes a chunk from `types.rs` and the corresponding entries
from the `lib.rs` flat re-export.
